Added form create an experiment by uploading a json with spectra and gensim model archive files
Had to move methods from run_gensim to load_dict_functions so they could be used in the celery task.
