+++ ms2ldaviz/templates/uploads/add_experiment.html
@@ -11,6 +11,7 @@
     <ul class="nav nav-tabs">
       <li role="presentation" class="active"><a href="/uploads/create_experiment">Create</a></li>
       <li role="presentation"><a href="/uploads/upload_experiment">Upload</a></li>
+      <li role="presentation"><a href="/uploads/upload_gensim_experiment">Upload gensim</a></li>
     </ul>
     <div class="panel-group">
         <div class="panel panel-default">
