+++ ms2ldaviz/uploads/urls.py
@@ -7,4 +7,5 @@ urlpatterns = [
     url(r'^process/$', views.process_experiment, name='process_experiment'),
     url(r'^ms1_format/$', TemplateView.as_view(template_name="uploads/ms1_format.html"), name='ms1_format'),
     url(r'^upload_experiment/$', views.upload_experiment, name='upload_experiment'),
+    url(r'^upload_gensim_experiment/$', views.upload_gensim_experiment, name='upload_gensim_experiment'),
 ]
