+++ ms2ldaviz/uploads/views.py
@@ -5,8 +5,8 @@ from django.shortcuts import render
 
 from basicviz.constants import EXPERIMENT_STATUS_CODE, EXPERIMENT_TYPE, EXPERIMENT_MS2_FORMAT
 from basicviz.models import UserExperiment,JobLog
-from .forms import CreateExperimentForm, UploadExperimentForm
-from .tasks import lda_task, decomposition_task, upload_task
+from .forms import CreateExperimentForm, UploadExperimentForm, UploadGensimExperimentForm
+from .tasks import lda_task, decomposition_task, upload_task, upload_gensim_task
 
 
 @login_required(login_url='/registration/login/')
@@ -68,6 +68,35 @@ def upload_experiment(request):
     return render(request, 'uploads/upload_experiment.html', context_dict)
 
 
+@login_required(login_url='/registration/login/')
+def upload_gensim_experiment(request):
+
+    context_dict = {}
+    if request.method == 'POST':
+        experiment_form = UploadGensimExperimentForm(request.POST, request.FILES)
+        if experiment_form.is_valid():
+            new_experiment = experiment_form.save(commit=False)
+            new_experiment.status = EXPERIMENT_STATUS_CODE[0][0]
+            new_experiment.experiment_ms2_format = EXPERIMENT_MS2_FORMAT[4][0]
+            new_experiment.ms2_file = request.FILES.get('ms2_file', None)
+            new_experiment.csv_file = request.FILES.get('csv_file', None)
+            new_experiment.save()
+
+            user = request.user
+            UserExperiment.objects.create(user=user, experiment=new_experiment, permission='edit')
+
+            JobLog.objects.create(user = user, experiment = new_experiment, tasktype = 'Uploaded and ' + new_experiment.experiment_type)
+            process_experiment(new_experiment, experiment_form.cleaned_data)
+            return HttpResponseRedirect(reverse('index'))
+        else:
+            context_dict['experiment_form'] = experiment_form
+    else:
+        experiment_form = UploadGensimExperimentForm()
+        context_dict['experiment_form'] = experiment_form
+
+    return render(request, 'uploads/upload_gensim_experiment.html', context_dict)
+
+
 def process_experiment(exp, cleaned_data):
 
     pending, desc = EXPERIMENT_STATUS_CODE[0]
@@ -84,6 +113,11 @@ def process_experiment(exp, cleaned_data):
                 params['filename'] = exp.ms2_file.path
                 params['featureset'] = exp.featureset.name
                 task = upload_task
+            elif exp.experiment_ms2_format == EXPERIMENT_MS2_FORMAT[4][0]:
+                params['corpus_filename'] = exp.ms2_file.path
+                params['gensim_filename'] = exp.csv_file.path
+                params['featureset'] = exp.featureset.name
+                task = upload_gensim_task
             else: # run lda inference
                 params['K'] = cleaned_data['K'] if 'K' in cleaned_data else None
                 params['n_its'] = cleaned_data['n_its'] if 'n_its' in cleaned_data else None
